refactor: create BddOpenai::FileClient to replace BddOpenai::Files::Client
