+++ lib/bdd_openai.rb
@@ -5,6 +5,6 @@ require 'net/http'
 require 'securerandom'
 require_relative 'bdd_openai/clients/http'
 require_relative 'bdd_openai/error_response'
-require_relative 'bdd_openai/files/client'
+require_relative 'bdd_openai/file_client'
 require_relative 'bdd_openai/mappers/file'
 require_relative 'bdd_openai/version'
