+++ lib/bdd_openai/clients/http.rb
@@ -5,26 +5,26 @@ module BddOpenai
   module Client
     # An HTTP client
     class HttpClient
-      # @param uri [String]
+      # @param uri [URI::Generic]
       # @param headers [Hash]
       # @param disable_ssl [Boolean]
       # @return [Net::HTTPResponse]
       def call_get(uri, headers, disable_ssl: false)
-        http = Net::HTTP.new(uri.host, uri.port)
+        http = Net::HTTP.new(uri.host || '', uri.port)
         http.use_ssl = disable_ssl ? false : true
-        request = Net::HTTP::Get.new(uri.path, headers)
+        request = Net::HTTP::Get.new(uri.path || '', headers)
         http.request(request)
       end
 
-      # @param uri [String]
+      # @param uri [URI::Generic]
       # @param req_body [String] Request body in the form of JSON string.
       # @param headers [Hash]
       # @param disable_ssl [Boolean]
       # @return [Net::HTTPResponse]
       def call_post(uri, req_body, headers, disable_ssl: false)
-        http = Net::HTTP.new(uri.host, uri.port)
+        http = Net::HTTP.new(uri.host || '', uri.port)
         http.use_ssl = disable_ssl ? false : true
-        request = Net::HTTP::Post.new(uri.path, headers)
+        request = Net::HTTP::Post.new(uri.path || '', headers)
         request.body = req_body unless req_body.nil?
         http.request(request)
       end
@@ -50,14 +50,14 @@ module BddOpenai
         [body, boundary]
       end
 
-      # @param uri [String]
+      # @param uri [URI::Generic]
       # @param headers [Hash]
       # @param disable_ssl [Boolean]
       # @return [Net::HTTPResponse]
       def call_delete(uri, headers, disable_ssl: false)
-        http = Net::HTTP.new(uri.host, uri.port)
+        http = Net::HTTP.new(uri.host || '', uri.port)
         http.use_ssl = disable_ssl ? false : true
-        request = Net::HTTP::Delete.new(uri.path, headers)
+        request = Net::HTTP::Delete.new(uri.path || '', headers)
         http.request(request)
       end
     end
